try a drupal 10 compatible trait with a hack
